Added average stay duration, updated to return double insted of float

+++ src/Implementations/MockQueryManager.java
@@ -50,7 +50,7 @@ public class MockQueryManager implements IQueryManager{
 		return customers;
 	}
 	
-	public int getAverageStayDuration(Calendar startDate, Calendar endDate) {
+	public double getAverageStayDuration(Calendar startDate, Calendar endDate) {
 		return 2;
 	}
